# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for 대출드림 loan platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Application {
  id: number;
  name: string;
  phone: string;
  job_type: string;
  /** @nullable */
  loan_amount?: string | null;
  /** @nullable */
  loan_purpose?: string | null;
  /** @nullable */
  residence_type?: string | null;
  /** @nullable */
  annual_income?: string | null;
  /** @nullable */
  credit_score?: string | null;
  /** @nullable */
  message?: string | null;
  status?: string;
  created_at: string;
}

export interface ApplicationInput {
  name: string;
  phone: string;
  job_type: string;
  loan_amount?: string;
  loan_purpose?: string;
  residence_type?: string;
  annual_income?: string;
  credit_score?: string;
  message?: string;
}

export interface ApplicationListResult {
  applications: Application[];
  total: number;
  page: number;
  limit: number;
}

export interface DeleteResult {
  success: boolean;
}

export interface AdminLoginInput {
  password: string;
}

export interface AdminLoginResult {
  token: string;
  success: boolean;
}

export interface JobTypeStat {
  job_type: string;
  count: number;
}

export interface AdminStats {
  total: number;
  today: number;
  by_job_type: JobTypeStat[];
  recent: Application[];
}

export interface TelegramSettings {
  enabled: boolean;
  /** @nullable */
  bot_token?: string | null;
  /** @nullable */
  chat_id?: string | null;
  /** @nullable */
  chat_name?: string | null;
}

export interface TelegramSettingsInput {
  enabled: boolean;
  bot_token?: string;
  chat_id?: string;
  chat_name?: string;
}

export interface TelegramDiscoverInput {
  bot_token: string;
}

export interface TelegramChat {
  id: string;
  name: string;
  type: string;
}

export interface TelegramDiscoverResult {
  chats: TelegramChat[];
}

export interface TelegramTestResult {
  success: boolean;
  message: string;
}

export type GetAdminApplicationsParams = {
page?: number;
limit?: number;
job_type?: string;
};