# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for 대출드림 loan platform
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: applications
    description: Loan application operations
  - name: admin
    description: Admin operations
  - name: telegram
    description: Telegram bot settings
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /applications:
    post:
      operationId: submitApplication
      tags: [applications]
      summary: Submit a loan application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationInput"
      responses:
        "201":
          description: Application submitted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "400":
          description: Invalid input

  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResult"
        "401":
          description: Unauthorized

  /admin/applications:
    get:
      operationId: getAdminApplications
      tags: [admin]
      summary: Get all loan applications (admin only)
      parameters:
        - name: x-admin-token
          in: header
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: job_type
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of applications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationListResult"
        "401":
          description: Unauthorized

  /admin/applications/{id}:
    delete:
      operationId: deleteApplication
      tags: [admin]
      summary: Delete an application (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: x-admin-token
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
        "401":
          description: Unauthorized
        "404":
          description: Not found

  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get admin dashboard statistics
      parameters:
        - name: x-admin-token
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
        "401":
          description: Unauthorized

  /admin/telegram:
    get:
      operationId: getTelegramSettings
      tags: [telegram]
      summary: Get Telegram bot settings
      parameters:
        - name: x-admin-token
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Telegram settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TelegramSettings"
        "401":
          description: Unauthorized
    put:
      operationId: updateTelegramSettings
      tags: [telegram]
      summary: Update Telegram bot settings
      parameters:
        - name: x-admin-token
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TelegramSettingsInput"
      responses:
        "200":
          description: Settings updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TelegramSettings"
        "401":
          description: Unauthorized

  /admin/telegram/discover:
    post:
      operationId: discoverTelegramChats
      tags: [telegram]
      summary: Discover chat rooms from bot token
      parameters:
        - name: x-admin-token
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TelegramDiscoverInput"
      responses:
        "200":
          description: Discovered chats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TelegramDiscoverResult"
        "400":
          description: Invalid token or no chats found
        "401":
          description: Unauthorized

  /admin/telegram/test:
    post:
      operationId: testTelegramNotification
      tags: [telegram]
      summary: Send a test Telegram notification
      parameters:
        - name: x-admin-token
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Test notification sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TelegramTestResult"
        "400":
          description: Telegram not configured
        "401":
          description: Unauthorized

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Application:
      type: object
      required: [id, name, phone, job_type, created_at]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        job_type:
          type: string
        loan_amount:
          type: ["string", "null"]
        loan_purpose:
          type: ["string", "null"]
        residence_type:
          type: ["string", "null"]
        annual_income:
          type: ["string", "null"]
        credit_score:
          type: ["string", "null"]
        message:
          type: ["string", "null"]
        status:
          type: string
        created_at:
          type: string

    ApplicationInput:
      type: object
      required: [name, phone, job_type]
      properties:
        name:
          type: string
        phone:
          type: string
        job_type:
          type: string
        loan_amount:
          type: string
        loan_purpose:
          type: string
        residence_type:
          type: string
        annual_income:
          type: string
        credit_score:
          type: string
        message:
          type: string

    ApplicationListResult:
      type: object
      required: [applications, total, page, limit]
      properties:
        applications:
          type: array
          items:
            $ref: "#/components/schemas/Application"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    DeleteResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean

    AdminLoginInput:
      type: object
      required: [password]
      properties:
        password:
          type: string

    AdminLoginResult:
      type: object
      required: [token, success]
      properties:
        token:
          type: string
        success:
          type: boolean

    AdminStats:
      type: object
      required: [total, today, by_job_type, recent]
      properties:
        total:
          type: integer
        today:
          type: integer
        by_job_type:
          type: array
          items:
            $ref: "#/components/schemas/JobTypeStat"
        recent:
          type: array
          items:
            $ref: "#/components/schemas/Application"

    JobTypeStat:
      type: object
      required: [job_type, count]
      properties:
        job_type:
          type: string
        count:
          type: integer

    TelegramSettings:
      type: object
      required: [enabled]
      properties:
        enabled:
          type: boolean
        bot_token:
          type: ["string", "null"]
        chat_id:
          type: ["string", "null"]
        chat_name:
          type: ["string", "null"]

    TelegramSettingsInput:
      type: object
      required: [enabled]
      properties:
        enabled:
          type: boolean
        bot_token:
          type: string
        chat_id:
          type: string
        chat_name:
          type: string

    TelegramDiscoverInput:
      type: object
      required: [bot_token]
      properties:
        bot_token:
          type: string

    TelegramChat:
      type: object
      required: [id, name, type]
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string

    TelegramDiscoverResult:
      type: object
      required: [chats]
      properties:
        chats:
          type: array
          items:
            $ref: "#/components/schemas/TelegramChat"

    TelegramTestResult:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string