# lib/api-zod/src/generated/types/adminStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for 대출드림 loan platform
 * OpenAPI spec version: 0.1.0
 */
import type { Application } from './application';
import type { JobTypeStat } from './jobTypeStat';

export interface AdminStats {
  total: number;
  today: number;
  by_job_type: JobTypeStat[];
  recent: Application[];
}