# lib/api-zod/src/generated/types/applicationInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for 대출드림 loan platform
 * OpenAPI spec version: 0.1.0
 */

export interface ApplicationInput {
  name: string;
  phone: string;
  job_type: string;
  loan_amount?: string;
  loan_purpose?: string;
  residence_type?: string;
  annual_income?: string;
  credit_score?: string;
  message?: string;
}