# lib/api-zod/src/generated/types/applicationListResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for 대출드림 loan platform
 * OpenAPI spec version: 0.1.0
 */
import type { Application } from './application';

export interface ApplicationListResult {
  applications: Application[];
  total: number;
  page: number;
  limit: number;
}