# lib/api-zod/src/generated/types/telegramSettingsInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for 대출드림 loan platform
 * OpenAPI spec version: 0.1.0
 */

export interface TelegramSettingsInput {
  enabled: boolean;
  bot_token?: string;
  chat_id?: string;
  chat_name?: string;
}